Moving logic for displaying actions into TargetedDataPanel.
