+++ src/summary/NarrativeNameValuePairsVisualizer.jsx
@@ -1,6 +1,6 @@
-import { ListItemIcon, ListItemText } from 'material-ui/List';
-import Menu, { MenuItem } from 'material-ui/Menu';
-import FontAwesome from 'react-fontawesome';
+//import { ListItemIcon, ListItemText } from 'material-ui/List';
+//import Menu, { MenuItem } from 'material-ui/Menu';
+//import FontAwesome from 'react-fontawesome';
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import Lang from 'lodash';
@@ -211,12 +211,13 @@ class NarrativeNameValuePairsVisualizer extends Component {
 
     // renders Menu for snippet and associated actions as Menu items
     // Will check whether an action should be rendered as a Menu item based on criteria of each action
-    renderedMenu = (snippet, snippetId, snippetText) => {
+    renderedMenu = (snippet, snippetId, snippetText, arrayIndex) => {
         const {
             snippetDisplayingMenu,
             positionLeft,
             positionTop,
         } = this.state;
+        
         const onMenuItemClicked = (fn, element) => {
             const callback = () => {
                 // convert element to format action is expecting
@@ -232,48 +233,13 @@ class NarrativeNameValuePairsVisualizer extends Component {
             }
             this.closeInsertionMenu(callback);
         }
-
+        
         let isSigned = true;
         const checkSnippetUnsigned = Lang.isUndefined(snippet.unsigned) ? isSigned : !snippet.unsigned;
         isSigned = Lang.isArray(snippet.value) ? !snippet.value[1] : checkSnippetUnsigned;
-        // Filter actions by whenToDisplay property on action
-        const filteredActions = this.props.actions.filter((a) => {
-            if (a.whenToDisplay.valueExists && Lang.isNull(snippet)) return false;
-            if (a.whenToDisplay.existingValueSigned !== "either" && a.whenToDisplay.existingValueSigned !== isSigned) return false;
-            return a.whenToDisplay.editableNoteOpen === "either" || String(a.whenToDisplay.editableNoteOpen) === String(this.props.allowItemClick);
-        });
-        if (filteredActions.length === 0) return null;
-        return (
-            <Menu
-                open={snippetDisplayingMenu === snippetId}
-                anchorReference="anchorPosition"
-                anchorPosition={{ top: positionTop, left: positionLeft }}
-                onClose={(event) => this.closeInsertionMenu()}
-                className="narrative-inserter-tooltip"
-            >
-                {
-                    // map filterActions to MenuItems
-                    filteredActions.map((a, index) => {
-                        const icon = a.icon ? (
-                            <ListItemIcon>
-                                <FontAwesome name={a.icon} />
-                            </ListItemIcon>
-                        ) : null;
-                        const text = a.text.replace("{elementText}", snippetText);
-                        return (
-                            <MenuItem
-                                key={`${snippetId}-${index}`}
-                                onClick={() => onMenuItemClicked(a.handler, snippet)}
-                                className="narrative-inserter-box"
-                            >
-                                {icon}
-                                <ListItemText className='narrative-inserter-menu-item' inset primary={text} />
-                            </MenuItem>
-                        )
-                    })
-                }
-            </Menu>
-        );
+     
+        return this.props.filterAndDisplayActions(this.props.actions, isSigned, arrayIndex, snippetId, snippet, snippetText, 
+                                                 this.closeInsertionMenu, onMenuItemClicked,  snippetDisplayingMenu, positionLeft, positionTop);
     }
 
     // Opens the insertion menu for the given snippet id, based on cursor location
@@ -321,7 +287,7 @@ class NarrativeNameValuePairsVisualizer extends Component {
                         >
                             {snippet.text}
                         </span>
-                        {this.renderedMenu(snippet.item, snippetId, snippet.text)}
+                        {this.renderedMenu(snippet.item, snippetId, snippet.text, index)}
                     </span>
                 );
             } else if (snippet.type !== 'plain') {
@@ -345,6 +311,7 @@ NarrativeNameValuePairsVisualizer.propTypes = {
     condition: PropTypes.object,
     conditionSection: PropTypes.object,
     isWide: PropTypes.bool,
+    filterAndDisplayActions: PropTypes.func.isRequired,
     allowItemClick: PropTypes.bool,
     actions: PropTypes.array
 };
